backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pymysql==1.1.1
bcrypt==4.1.2
PyJWT==2.9.0
scikit-learn==1.5.2
pandas==2.2.3
joblib==1.4.2
python-multipart==0.0.9
requests==2.32.3
python-docx==1.1.2
pdfplumber==0.11.4
openai==1.54.3
google-generativeai==0.8.3